requirements.txt
google.genai 
Streamlit 
quiz_generator